Create app skeleton #01 (#1)
* style: Add Bitter and Montserrat fonts to stylesheet

Create a folder for the stylesheet called Style. Within the folder a style.css
was created and modified with an import for the Google Fonts Bitter and Montserrat.

Resolves: #01

* style: Add reset default browser styles

Add normalize.css to normalize all default css across all browser css presets.

Resolves: #01

* feat: Add folder Components and Search.js within

The start of the components folder that will hold all components
and routes for the application. Within the first route Search.js
was created with a basic template.

Resolves: #01

* feat: Add react-router-dom to package.json

react-router-dom will allow the application to route
between specific paths.

Resolves: #01

* feat: Add routes within the App.js

Create the basic template for routes within the App.js
file. This will be expanded upon later but for now a template
will suffice.

Resolves: #01

* style: Add global styles

First add of the global styles for the stylesheet
style.css

Resolves: #01

* style: Refactor normalize.css to single location

normalize.css did not need its own css file and instead could
be imported in the style.css file within a single line.

Resolves: #01

* style: Add fonts the right way

This is the proper way to download the fonts since
we will be using styled-components rather than a css.

Resolves: #01

* feat: Add the propery dependecies for setup

Add dependecies for react-router-dom, styled-components,
and styled-normalize.

Resolves: #01

* style: Remove of style.css

Remove of style.css since we will be using
styled-components instead.

Resolves: #01

* refactor: Remove of base react images

Removal of the standard react images with an update to
the pages description to better match what the site is
actually about.

Resolves: #01

* refactor: Update Routes found in App.js

Update App.js to have better readability for the routes
by making them on a single line rather than span across
three lines. Also Normalize is refactored to be imported
from the dependecy rather than from a css file.

Resolves: #01

* style: Add global style to app

This is the proper way we should have added global
styles to the app with styled-components rather than
use a css file.

Resolves: #01

* style: Add theme to App.js

Create theme within the theme.js and imported
it within the App.js and wrapped everything within
ThemeProvider.

Resolves: #01